moving many things much
