+++ libraries/action/src/commonMain/kotlin/com/zegreatrob/coupling/action/pairassignmentdocument/Clock.kt
@@ -1,4 +1,4 @@
-package com.zegreatrob.coupling.server.action.pairassignmentdocument
+package com.zegreatrob.coupling.action.pairassignmentdocument
 
 interface Clock {
     fun currentDate() = kotlinx.datetime.Clock.System.now()
