+++ libraries/action/src/commonMain/kotlin/com/zegreatrob/coupling/action/pairassignmentdocument/Round.kt
@@ -1,0 +1,6 @@
+package com.zegreatrob.coupling.action.pairassignmentdocument
+
+import com.zegreatrob.coupling.model.pairassignmentdocument.CouplingPair
+import kotools.types.collection.NotEmptyList
+
+data class Round(val pairs: NotEmptyList<CouplingPair>, val gameSpin: GameSpin?)
