+++ libraries/action/src/jsMain/kotlin/com/zegreatrob/coupling/action/pairassignmentdocument/CreatePairCandidateReportAction.kt
@@ -1,4 +1,4 @@
-package com.zegreatrob.coupling.server.action.pairassignmentdocument
+package com.zegreatrob.coupling.action.pairassignmentdocument
 
 import com.zegreatrob.coupling.model.pairassignmentdocument.NeverPaired
 import com.zegreatrob.coupling.model.pairassignmentdocument.PairAssignmentDocument
