+++ server/actionz/src/jsMain/kotlin/com/zegreatrob/coupling/server/action/connection/ConnectPartyUserCommand.kt
@@ -1,12 +1,12 @@
 package com.zegreatrob.coupling.server.action.connection
 
+import com.zegreatrob.coupling.action.CannonProvider
 import com.zegreatrob.coupling.action.valueOrNull
 import com.zegreatrob.coupling.model.CouplingConnection
 import com.zegreatrob.coupling.model.party.PartyId
 import com.zegreatrob.coupling.model.player.Player
 import com.zegreatrob.coupling.model.player.matches
 import com.zegreatrob.coupling.model.user.CurrentUserProvider
-import com.zegreatrob.coupling.server.action.CannonProvider
 import com.zegreatrob.coupling.server.action.user.UserIsAuthorizedWithDataAction
 import com.zegreatrob.testmints.action.annotation.ActionMint
