+++ server/actionz/src/jsTest/kotlin/com/zegreatrob/coupling/server/action/pairassignmentdocument/CreatePairCandidateReportsActionTest.kt
@@ -1,5 +1,10 @@
 package com.zegreatrob.coupling.server.action.pairassignmentdocument
 
+import com.zegreatrob.coupling.action.pairassignmentdocument.CreatePairCandidateReportAction
+import com.zegreatrob.coupling.action.pairassignmentdocument.CreatePairCandidateReportListAction
+import com.zegreatrob.coupling.action.pairassignmentdocument.GameSpin
+import com.zegreatrob.coupling.action.pairassignmentdocument.PairCandidateReport
+import com.zegreatrob.coupling.action.pairassignmentdocument.let
 import com.zegreatrob.coupling.model.forEach
 import com.zegreatrob.coupling.model.pairassignmentdocument.NeverPaired
 import com.zegreatrob.coupling.model.pairassignmentdocument.PairAssignmentDocument
