+++ server/src/jsMain/kotlin/com/zegreatrob/coupling/server/entity/pairassignment/PairAssignmentDispatcher.kt
@@ -1,15 +1,15 @@
 package com.zegreatrob.coupling.server.entity.pairassignment
 
 import com.zegreatrob.coupling.action.pairassignmentdocument.AssignPinsAction
-import com.zegreatrob.coupling.server.action.pairassignmentdocument.CreatePairCandidateReportAction
-import com.zegreatrob.coupling.server.action.pairassignmentdocument.CreatePairCandidateReportListAction
-import com.zegreatrob.coupling.server.action.pairassignmentdocument.FindNewPairsAction
-import com.zegreatrob.coupling.server.action.pairassignmentdocument.NextPlayerAction
-import com.zegreatrob.coupling.server.action.pairassignmentdocument.ShufflePairsAction
-import com.zegreatrob.coupling.server.action.pairassignmentdocument.Wheel
+import com.zegreatrob.coupling.action.pairassignmentdocument.CreatePairCandidateReportAction
+import com.zegreatrob.coupling.action.pairassignmentdocument.CreatePairCandidateReportListAction
+import com.zegreatrob.coupling.action.pairassignmentdocument.FindNewPairsAction
+import com.zegreatrob.coupling.action.pairassignmentdocument.NextPlayerAction
+import com.zegreatrob.coupling.action.pairassignmentdocument.ShufflePairsAction
+import com.zegreatrob.coupling.action.pairassignmentdocument.Wheel
 
 interface PairAssignmentDispatcher<D> :
-    ShufflePairsAction.Dispatcher<D>,
+    com.zegreatrob.coupling.action.pairassignmentdocument.ShufflePairsAction.Dispatcher<D>,
     FindNewPairsAction.Dispatcher<D>,
     NextPlayerAction.Dispatcher<D>,
     CreatePairCandidateReportListAction.Dispatcher<D>,
